+++ tests/test_fitting.py
@@ -191,105 +191,3 @@ class TestBeamfit(unittest.TestCase):
             np.testing.assert_allclose(
                 beamfit.gaussian_lls_trans_grad(x0), j, atol=1e-10
             )
-
-
-class TestAnalysisMethod(unittest.TestCase):
-    def test_threshold(self):
-        m, n = np.mgrid[:64, :64]
-        image = beamfit.supergaussian(n, m, 32, 32, 8**2, 0, 8**2, 1, 1, 0)
-        o = beamfit.AnalysisMethodDebugger(sigma_threshold=2).fit(image)
-        self.assertEqual(np.sum(o.mask), np.sum(image < np.exp(-(2**2))))
-
-    def test_median_filter(self):
-        m, n = np.mgrid[:64, :64]
-        image = beamfit.supergaussian(n, m, 32, 32, 8**2, 0, 8**2, 1, 1, 0)
-        image[4, 24] = 100  # Set some random hot pixels
-        image[32, 15] = 100
-        image[54, 22] = 100
-        o = beamfit.AnalysisMethodDebugger(median_filter_size=3).fit(image)
-        self.assertAlmostEqual(o.max(), 1, places=1)
-
-    def test_masked(self):
-        m, n = np.mgrid[:64, :64]
-        image = beamfit.supergaussian(n, m, 32, 32, 8**2, 0, 8**2, 1, 1, 0)
-        o = beamfit.AnalysisMethodDebugger(sigma_threshold=2, median_filter_size=3).fit(
-            image
-        )
-        self.assertTrue(np.ma.isMaskedArray(o))
-
-
-class TestSigmaParameterization(unittest.TestCase):
-    def setUp(self):
-        self.matrices = [
-            np.identity(2),
-            np.array([[1, 2], [2, 5]]),
-            np.array([[100, 1], [1, 0.1]]),
-        ]
-        self.m = np.array([[1, 1], [1, 5]])  # Example from paper I am following
-        self.ps = [
-            beamfit.Cholesky(),
-            beamfit.LogCholesky(),
-            beamfit.Spherical(),
-            beamfit.MatrixLogarithm(),
-            beamfit.Givens(),
-        ]
-
-    def test_inverses(self):
-        """Make sure all parameterizations have a valid inverse"""
-        for p in self.ps:
-            for m in self.matrices:
-                try:
-                    np.testing.assert_allclose(p.reverse(p.forward(m)), m, atol=1e-9)
-                except:
-                    print(f"Failed at {p}")
-                    raise
-
-    def test_cholesky(self):
-        np.testing.assert_allclose(
-            beamfit.Cholesky().forward(self.m), np.array([1, 1, 2]), atol=1e-9
-        )
-
-    def test_log_cholesky(self):
-        np.testing.assert_allclose(
-            beamfit.LogCholesky().forward(self.m),
-            np.array([0, 1, np.log(2)]),
-            atol=1e-9,
-        )
-
-    def test_spherical(self):
-        np.testing.assert_allclose(
-            beamfit.Spherical().forward(self.m),
-            np.array([0, np.log(5) / 2, -0.608]),
-            rtol=1e-3,
-        )
-
-    def test_grads_numerical(self):
-        x0 = np.array([2, 3, 5])
-        hs = np.ones(100) * 1e-4
-        hs[0] = 1.0  # Manually set h values for specific parameterizations
-
-        for p, h in zip(self.ps, hs):
-            try:
-                j, err = calc_gradient_central_difference(p.reverse, x0=x0, h=h)
-                # Next line gives estimate of relative truncation error, but doesn't include roundoff error. Try to
-                # select the largest h such that the error is like 1e-9.
-                # print(err)
-                self.assertTrue(
-                    (err[np.isfinite(err)] < 1e-6).all()
-                )  # If it fails here, need to make h smaller
-                j_actual = p.reverse_grad(x0)
-                np.testing.assert_allclose(j_actual, j, atol=1e-8, rtol=1e-5)
-            except:
-                print(f"Failed at {p}")
-                raise
-
-    def test_eigen2d_grad_numerical(self):
-        x0 = np.array([2, 3, 5])
-
-        def rev(s):
-            a = beamfit.eigen2d(s)
-            return np.array([[a[0], a[1]], [a[1], a[2]]])
-
-        j, err = calc_gradient_central_difference(rev, x0=x0, h=1e-4)
-        j_actual = beamfit.eigen2d_grad(x0)
-        np.testing.assert_allclose(j_actual, j, atol=1e-8, rtol=1e-5)
